Create route search

+++ src/RoutePlanning.Client.Web/Pages/DistanceCalculator.razor.cs
@@ -12,7 +12,7 @@ public sealed partial class DistanceCalculator
     private SelectableLocation? SelectedDestination { get; set; }
     private string? DisplaySource { get; set; }
     private string? DisplayDestination { get; set; }
-    private int? DisplayDistance { get; set; }
+    private double? DisplayDistance { get; set; }
 
     [Inject]
     private IMediator Mediator { get; set; } = default!;
